[Ref] Silence signed/unsigned comparison warning.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@3193 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ soundlib/Load_dbm.cpp
@@ -141,7 +141,7 @@ struct PACKED DBMEnvelope
 	void ConvertEndianness()
 	{
 		SwapBytesBE(instrument);
-		for(int i = 0; i < CountOf(data); i++)
+		for(size_t i = 0; i < CountOf(data); i++)
 		{
 			SwapBytesBE(data[i]);
 		}
